Writing search history.

+++ app/src/main/java/ru/brauer/cleanarchitecture/model/datasource/database/SearchWord.kt
@@ -6,6 +6,6 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "search_history")
 data class SearchWord(
-    @PrimaryKey @ColumnInfo(name = "data_time") val dateTime: Int,
+    @PrimaryKey @ColumnInfo(name = "data_time") val dateTime: Long,
     @ColumnInfo(name = "search_word") val searchWord: String
 )
